+++ tests/modeling/test_modeling.py
@@ -15,6 +15,7 @@ from pharmpy.modeling import (
     add_iov,
     add_lag_time,
     create_joint_distribution,
+    has_michaelis_menten_elimination,
     load_example_model,
     remove_iiv,
     remove_iov,
@@ -70,7 +71,9 @@ $ESTIMATION METHOD=1 INTERACTION
 
 def test_set_michaelis_menten_elimination(testdata):
     model = Model.create_model(testdata / 'nonmem' / 'pheno.mod')
+    assert not has_michaelis_menten_elimination(model)
     set_michaelis_menten_elimination(model)
+    assert has_michaelis_menten_elimination(model)
     correct = """$PROBLEM PHENOBARB SIMPLE MODEL
 $DATA pheno.dta IGNORE=@
 $INPUT ID TIME AMT WGT APGR DV
@@ -96,6 +99,8 @@ $SIGMA 0.013241
 
 $ESTIMATION METHOD=1 INTERACTION
 """
+    assert model.model_code == correct
+    set_michaelis_menten_elimination(model)
     assert model.model_code == correct