+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ next version
 
 * Add modeling.calculate_aic
 * Add modeling.print_model_code
+* Add modeling.has_michaelis_menten_elimination
 
 0.45.0 (2022-01-21)
 -------------------